rename `ValkyrieCollectionIndexer` to support name resolution
ensures `ValkyrieIndexer.for` resolves the correct indexer for collections.
