+++ app/indexers/hyrax/valkyrie_collection_indexer.rb
@@ -3,17 +3,8 @@
 module Hyrax
   ##
   # Indexes properties common to PCDM Collections
-  class ValkyrieCollectionIndexer < Hyrax::ValkyrieIndexer
-    include Hyrax::ResourceIndexer
-    include Hyrax::PermissionIndexer
-    include Hyrax::VisibilityIndexer
-    include Hyrax::Indexer(:core_metadata)
-
-    def to_solr
-      super.tap do |index_document|
-        index_document[:generic_type_sim] = ['Collection']
-        index_document[:thumbnail_path_ss] = Hyrax::CollectionThumbnailPathService.call(resource)
-      end
-    end
+  #
+  # @deprecated use Hyrax::PcdmCollectionIndexer instead
+  class ValkyrieCollectionIndexer < Hyrax::PcdmCollectionIndexer
   end
 end
